refactor create_df
